+++ src/main/kotlin/org/xbery/artbeams/common/access/service/UserAccessService.kt
@@ -1,8 +1,8 @@
 package org.xbery.artbeams.common.access.service
 
+import jakarta.servlet.http.HttpServletRequest
 import org.xbery.artbeams.common.access.domain.EntityKey
 import org.xbery.artbeams.common.access.domain.UserAccessReport
-import jakarta.servlet.http.HttpServletRequest
 
 /**
  * Operations for user accesses.
@@ -25,7 +25,7 @@ interface UserAccessService {
      */
     fun saveUserAccess(entityKey: EntityKey, request: HttpServletRequest): UserAccessReport
 
-    fun findCountOfVisits(entityKey: EntityKey): Long
+    fun findCountOfVisits(entityKey: EntityKey): Int
 
     fun aggregateUserAccesses(): Unit
 }
